+++ BankSystem.Database/Account.sql
@@ -3,6 +3,6 @@
 	[IBANNumber] NCHAR(18) NOT NULL PRIMARY KEY, 
     [Name] NVARCHAR(50) NOT NULL, 
     [IsActive] BIT NOT NULL DEFAULT 1, 
-    [Balance] DECIMAL(18, 2) NOT NULL DEFAULT 0.00, 
+    [Balance] FLOAT NOT NULL DEFAULT 0.00, 
     [CreatedDate] DATETIME NOT NULL
 )
